ChapsterInterface/requirements.txt
keyboard==0.13.5
paho_mqtt==1.6.1
